Split lib/util.js
