+++ lib/runners/bf.js
@@ -1,10 +1,12 @@
-var shovel = require('../shovel'),
-    util = require('../util');
+"use strict";
+
+const shovel = require('../shovel');
+const writeFileSync = require('../utils/write-file-sync');
 
 module.exports.run = function run(opts, cb) {
   shovel.start(opts, cb, {
     solutionOnly: function(runCode) {
-      var solutionFile = util.writeFileSync('/home/codewarrior', 'code.b', opts.solution, true);
+      const solutionFile = writeFileSync('/home/codewarrior', 'code.b', opts.solution, true);
       runCode({
         name: 'bf',
         args: ['-c29999', solutionFile]
