+++ lib/runners/elixir.js
@@ -1,6 +1,9 @@
-var shovel = require('../shovel'),
-    util = require('../util'),
-    temp = require('temp');
+"use strict";
+
+const temp = require('temp');
+
+const shovel = require('../shovel');
+const codeWriteSync = require('../utils/code-write-sync');
 
 module.exports.run = function run(opts, cb) {
   shovel.start(opts, cb, {
@@ -25,8 +28,8 @@ module.exports.run = function run(opts, cb) {
 
 function prepareExUnit(opts, run) {
   var dir = temp.mkdirSync('elixir');
-  var solutionFile = util.codeWriteSync('elixir', solution(opts), dir, 'solution', true);
-  var fixtureFile = util.codeWriteSync('elixir', opts.fixture, dir, 'fixture', true);
+  var solutionFile = codeWriteSync('elixir', solution(opts), dir, 'solution', true);
+  var fixtureFile = codeWriteSync('elixir', opts.fixture, dir, 'fixture', true);
 
   var code = `
         Code.load_file("frameworks/elixir/cw_runner.ex")
