+++ lib/runners/erlang.js
@@ -1,7 +1,7 @@
 "use strict";
 
 const shovel = require('../shovel');
-const codeWriteSync = require('../util').codeWriteSync;
+const codeWriteSync = require('../utils/code-write-sync');
 const path = require('path');
 const fs = require('fs');
 const temp = require('temp');
