+++ lib/runners/haskell.js
@@ -1,6 +1,9 @@
-var shovel = require('../shovel'),
-    codeWriteSync = require('../util').codeWriteSync,
-    temp = require('temp');
+"use strict";
+
+const temp = require('temp');
+
+const shovel = require('../shovel');
+const codeWriteSync = require('../utils/code-write-sync');
 
 module.exports.run = function run(opts, cb) {
   temp.track();
