+++ lib/runners/javascript.js
@@ -5,7 +5,7 @@ const path = require('path');
 const fs = require('fs-extra');
 
 const shovel = require('../shovel');
-const util = require('../util');
+const writeFilesSync = require('../utils/write-files-sync');
 const frontendFrameworks = require('../frontend-frameworks');
 
 
@@ -60,7 +60,7 @@ module.exports.run = function run(opts, cb) {
       // a files object/hash can be passed in as options, which will be transformed (if applicable) and then stored
       // in the provided directory.
       // write files that may have been included to the file system, within the directory that the code executes in
-      util.writeFilesSync(opts.dir, opts.files, true, function(name, content) {
+      writeFilesSync(opts.dir, opts.files, true, function(name, content) {
         return /\.js$/.test(name) ? maybeTransform(content, opts) : content;
       });
     }
