+++ lib/runners/julia.js
@@ -1,8 +1,9 @@
-'use strict';
+"use strict";
 
-var shovel = require('../shovel'),
-    codeWriteSync = require('../util').codeWriteSync,
-    temp = require('temp');
+const temp = require('temp');
+
+const shovel = require('../shovel');
+const codeWriteSync = require('../utils/code-write-sync');
 
 module.exports.run = function run(opts, cb) {
   temp.track();
