+++ lib/runners/lua.js
@@ -1,7 +1,7 @@
 "use strict";
 
 const shovel = require('../shovel');
-const writeFileSync = require('../util').writeFileSync;
+const writeFileSync = require('../utils/write-file-sync');
 
 module.exports.run = function run(opts, cb) {
   shovel.start(opts, cb, {
