+++ lib/runners/objc.js
@@ -1,7 +1,10 @@
-var shovel = require('../shovel'),
-    util = require('../util'),
-    path = require('path'),
-    exec = require('child_process').exec;
+"use strict";
+
+const path = require('path');
+const exec = require('child_process').exec;
+
+const codeWriteSync = require('../utils/code-write-sync');
+const shovel = require('../shovel');
 
 function getCode(opts) {
   if (opts.mode === "default") {
@@ -53,11 +56,11 @@ function prepareCW(opts) {
       }
   `;
 
-  util.codeWriteSync('objc', opts.solution, opts.dir, 'solution.m');
+  codeWriteSync('objc', opts.solution, opts.dir, 'solution.m');
   if (opts.setup) {
-    util.codeWriteSync('objc', opts.setup, opts.dir, 'setup.m');
+    codeWriteSync('objc', opts.setup, opts.dir, 'setup.m');
   }
-  return [util.codeWriteSync('objc', main, opts.dir, 'main.m')];
+  return [codeWriteSync('objc', main, opts.dir, 'main.m')];
 }
 
 function prepareUnitKit(opts) {
@@ -123,10 +126,10 @@ function prepareUnitKit(opts) {
       }
   `;
 
-  util.codeWriteSync('objc', getCode(opts), opts.dir, 'solution.m');
-  if (opts.setup) util.codeWriteSync('objc', opts.setup, opts.dir, 'setup.m');
-  const fixtureFile = util.codeWriteSync('objc', fixture, opts.dir, 'fixture.m');
-  const mainFile = util.codeWriteSync('objc', main, opts.dir, 'main.m');
+  codeWriteSync('objc', getCode(opts), opts.dir, 'solution.m');
+  if (opts.setup) codeWriteSync('objc', opts.setup, opts.dir, 'setup.m');
+  const fixtureFile = codeWriteSync('objc', fixture, opts.dir, 'fixture.m');
+  const mainFile = codeWriteSync('objc', main, opts.dir, 'main.m');
   return [mainFile, fixtureFile, '-lUnitKit'];
 }
 
@@ -156,7 +159,7 @@ module.exports.run = function run(opts, cb) {
 
       opts.solution = getCode(opts);
 
-      var solutionFile = util.codeWriteSync('objc', opts.solution, opts.dir, 'solution.m');
+      var solutionFile = codeWriteSync('objc', opts.solution, opts.dir, 'solution.m');
 
       args = ['-o', executable, solutionFile];
